fix(sprite-depth-spike): address final-review comment/doc/report findings
No runtime behavior change. Fixes from the whole-branch review:

- Rewrite the depth-clamp comment in sprite_depth_spike.wgsl: clamping a
  behind-camera fragment to depth 1.0 makes it incorrectly WIN the reverse-Z
  depth test, not "pin sanely" as previously claimed. Note M8 needs an
  explicit discard/clamp/constrain decision here.
- Fix depth_step_rgba's doc comment, which contradicted depth_step_image's
  (it claimed sRGB storage; depth_step_image is explicitly non-sRGB).
- Add the missing failure mode (per-pixel depth not reaching the depth
  buffer) to the interpenetration test's timeout assertion message.
- Reorder sprite_depth_spike's mod/use entries alphabetically in lib.rs.
- Match solid_white_image's usize-widening order to depth_step_rgba's.
- Reword the findings report's cache-coldness claim: both measured runs
  converged at 5 updates, which is more likely evidence of a persistent
  driver cache than a true cold-cache figure; point future readers at the
  plan's documented 60-update worst case instead. Add a "what M8 inherits"
  bullet for the depth-clamp finding.

+++ crates/sway-runtime/src/sprite_depth_spike.rs
@@ -113,7 +113,8 @@ impl Plugin for SpriteDepthPlugin {
 }
 
 /// A single-channel-in-R depth atlas: left half near (0.0), right half far
-/// (1.0). Stored `Rgba8UnormSrgb` for one reason — see `depth_step_image`.
+/// (1.0). Raw bytes consumed by `depth_step_image`, which deliberately does
+/// *not* interpret them as sRGB — see that function's doc comment for why.
 pub fn depth_step_rgba(size: u32) -> Vec<u8> {
     let mut data = Vec::with_capacity((size as usize) * (size as usize) * 4);
     for _y in 0..size {
@@ -146,7 +147,7 @@ pub fn solid_white_image(size: u32) -> Image {
     Image::new(
         Extent3d { width: size, height: size, depth_or_array_layers: 1 },
         TextureDimension::D2,
-        vec![255u8; (size * size * 4) as usize],
+        vec![255u8; (size as usize) * (size as usize) * 4],
         TextureFormat::Rgba8UnormSrgb,
         RenderAssetUsages::RENDER_WORLD,
     )
